+++ src/scanner/ethereum_uniswap.cpp
@@ -1,0 +1,11 @@
+#include "functions/scanner.hpp"
+#include "duckdb/common/helper.hpp"
+#include <curl/curl.h>
+#include <iostream>
+#include <sstream>
+#include <string>
+#include "json.hpp"
+
+namespace duckdb {
+namespace scrooge {}
+} // namespace duckdb
